new protocol header implementation for typed requests in memcache
Summary: This diff includes Header parsing for the new protocol(Caret ^) for typed requests in Umbrella.

Test Plan: tests

Reviewed By: @alikhtarov

Differential Revision: D2306204

+++ mcrouter/lib/network/McServerSession.cpp
@@ -295,7 +295,7 @@ void McServerSession::requestReady(McRequest&& req,
   }
 }
 
-void McServerSession::typedRequestReady(uint64_t typeId,
+void McServerSession::typedRequestReady(uint32_t typeId,
                                         const folly::IOBuf& reqBody,
                                         uint64_t reqid) {
   DestructorGuard dg(this);
